chore: mod.tsでオブジェクトを作り，intermoduleで再公開する
