+++ pkg/intermodule/note.ts
@@ -1,35 +1,23 @@
-import { Cat, Ether, type Result } from '@mikuroxina/mini-fn';
+import { Ether, type Result } from '@mikuroxina/mini-fn';
 import type { AccountID } from '../accounts/model/account.js';
-import { isProduction } from '../adaptors/env.js';
-import { prismaClient } from '../adaptors/prisma.js';
 import type { Medium } from '../drive/model/medium.js';
-import { clockSymbol, snowflakeIDGenerator } from '../id/mod.js';
 import {
-  inMemoryNoteAttachmentRepo,
-  inMemoryNoteRepo,
-  inMemoryReactionRepo,
-} from '../notes/adaptor/repository/dummy.js';
-import {
-  prismaNoteAttachmentRepo,
-  prismaNoteRepo,
-  prismaReactionRepo,
-} from '../notes/adaptor/repository/prisma.js';
+  noteAttachmentRepoEther,
+  noteClockEther,
+  noteCreateServiceInstance,
+  noteFetchServiceInstance,
+  noteIdGeneratorEther,
+  noteReactionRepoEther,
+  noteRepoEther,
+} from '../notes/mod.js';
 import type { Note, NoteID } from '../notes/model/note.js';
 import type { Reaction } from '../notes/model/reaction.js';
 import type { RenoteStatus } from '../notes/model/renoteStatus.js';
-import { type CreateService, createService } from '../notes/service/create.js';
-import { type FetchService, fetch } from '../notes/service/fetch.js';
-import {
-  accountModule,
-  accountModuleFacadeSymbol,
-  dummyAccountModuleFacade,
-} from './account.js';
-import { timelineModuleFacadeEther } from './timeline.js';
 
 export class NoteModuleFacade {
   constructor(
-    private readonly fetchService: FetchService,
-    private readonly createNoteService: CreateService,
+    private readonly fetchService: typeof noteFetchServiceInstance,
+    private readonly createNoteService: typeof noteCreateServiceInstance,
   ) {}
 
   /**
@@ -74,51 +62,18 @@ export class NoteModuleFacade {
   }
 }
 
-// NOTE: These dependency Ethers are shared between intermodule and notes module to ensure the same instances are used
-export const noteAttachmentRepoEther = isProduction
-  ? prismaNoteAttachmentRepo(prismaClient)
-  : inMemoryNoteAttachmentRepo([], []);
-export const noteReactionRepoEther = isProduction
-  ? prismaReactionRepo(prismaClient)
-  : inMemoryReactionRepo([]);
-export const noteRepoEther = isProduction
-  ? prismaNoteRepo(prismaClient)
-  : inMemoryNoteRepo([]);
-const accountModuleFacade = Ether.newEther(accountModuleFacadeSymbol, () =>
-  isProduction ? accountModule : dummyAccountModuleFacade,
-);
+// NOTE: Re-export dependency Ethers and Service instances from notes module
+export {
+  noteAttachmentRepoEther,
+  noteClockEther,
+  noteCreateServiceInstance,
+  noteFetchServiceInstance,
+  noteIdGeneratorEther,
+  noteReactionRepoEther,
+  noteRepoEther,
+};
+
 export const noteModuleFacadeSymbol = Ether.newEtherSymbol<NoteModuleFacade>();
-
-class Clock {
-  now() {
-    return BigInt(Date.now());
-  }
-}
-export const noteClockEther = Ether.newEther(clockSymbol, () => new Clock());
-export const noteIdGeneratorEther = Ether.compose(noteClockEther)(
-  snowflakeIDGenerator(0),
-);
-
-/**
- *  Shared Service Instances
- *  NOTE: These instances are shared between intermodule and notes module to ensure subscription works correctly
- */
-export const noteFetchServiceInstance = Ether.runEther(
-  Cat.cat(fetch)
-    .feed(Ether.compose(noteRepoEther))
-    .feed(Ether.compose(accountModuleFacade))
-    .feed(Ether.compose(noteAttachmentRepoEther))
-    .feed(Ether.compose(noteReactionRepoEther)).value,
-);
-
-export const noteCreateServiceInstance = Ether.runEther(
-  Cat.cat(createService)
-    .feed(Ether.compose(noteRepoEther))
-    .feed(Ether.compose(noteClockEther))
-    .feed(Ether.compose(noteIdGeneratorEther))
-    .feed(Ether.compose(noteAttachmentRepoEther))
-    .feed(Ether.compose(timelineModuleFacadeEther)).value,
-);
 
 /**
  *  Dependency Injected NoteModule Object
